Closes #8, closes #84
